# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KOReader Smart Library + Progress Sync API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BookCard {
  md5: string;
  name: string;
  display_title: string;
  display_author: string;
  /** @nullable */
  path?: string | null;
  ext: string;
  folder: string;
  size_human: string;
  mtime_iso: string;
  has_cover: boolean;
  last_progress: number;
  last_read_iso: string;
  last_user: string;
  last_device: string;
  last_ts: number;
  epub_has_meta: boolean;
  epub_title: string;
  epub_authors: string[];
  epub_publisher: string;
  epub_language: string;
  epub_date: string;
  epub_series: string;
  epub_series_index: string;
  epub_subjects: string[];
  epub_identifiers: string[];
  epub_description: string;
}

export interface UserStat {
  user: string;
  books: number;
  last_ts: number;
  last_iso: string;
  devices: string[];
}

export interface StatsResponse {
  active_books: number;
  completed: number;
  in_progress: number;
  not_started: number;
  avg_progress: number;
  users: number;
  last_activity_iso: string;
  last_activity_age: string;
  days_active_last14: number;
  top_recent: BookCard[];
  users_breakdown: UserStat[];
}

export interface SyncProgress {
  document: string;
  progress: string;
  percentage: number;
  device: string;
  device_id: string;
  timestamp: number;
}

export interface SyncProgressUpdate {
  progress: string;
  percentage: number;
  device: string;
  device_id: string;
}

export interface AuthResponse {
  authorized: boolean;
  user: string;
}

export type SearchLibraryParams = {
  /**
   * Search query
   */
  q: string;
  /**
   * Filter by file extension (epub, pdf, etc)
   */
  ext?: string;
  /**
   * Set to "1" to only return books with covers
   */
  cover?: string;
  /**
   * Set to "1" to only return books with reading progress
   */
  recent?: string;
};