# lib/api-zod/src/generated/types/bookCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KOReader Smart Library + Progress Sync API
 * OpenAPI spec version: 0.1.0
 */

export interface BookCard {
  md5: string;
  name: string;
  display_title: string;
  display_author: string;
  /** @nullable */
  path?: string | null;
  ext: string;
  folder: string;
  size_human: string;
  mtime_iso: string;
  has_cover: boolean;
  last_progress: number;
  last_read_iso: string;
  last_user: string;
  last_device: string;
  last_ts: number;
  epub_has_meta: boolean;
  epub_title: string;
  epub_authors: string[];
  epub_publisher: string;
  epub_language: string;
  epub_date: string;
  epub_series: string;
  epub_series_index: string;
  epub_subjects: string[];
  epub_identifiers: string[];
  epub_description: string;
}