# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KOReader Smart Library + Progress Sync API
 * OpenAPI spec version: 0.1.0
 */
import type { BookCard } from "./bookCard";
import type { UserStat } from "./userStat";

export interface StatsResponse {
  active_books: number;
  completed: number;
  in_progress: number;
  not_started: number;
  avg_progress: number;
  users: number;
  last_activity_iso: string;
  last_activity_age: string;
  days_active_last14: number;
  top_recent: BookCard[];
  users_breakdown: UserStat[];
}