# lib/api-zod/src/generated/types/syncProgressUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KOReader Smart Library + Progress Sync API
 * OpenAPI spec version: 0.1.0
 */

export interface SyncProgressUpdate {
  progress: string;
  percentage: number;
  device: string;
  device_id: string;
}