# lib/api-zod/src/generated/types/userStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KOReader Smart Library + Progress Sync API
 * OpenAPI spec version: 0.1.0
 */

export interface UserStat {
  user: string;
  books: number;
  last_ts: number;
  last_iso: string;
  devices: string[];
}